Split built-in theme definitions
